requirements.txt
certifi==2017.4.17
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
Django==1.11.2
django-crispy-forms==1.6.1
django-filter==1.0.4
django-guardian==1.4.8
djangorestframework==3.6.3
idna==2.5
itypes==1.1.0
Jinja2==2.9.6
Markdown==2.6.8
MarkupSafe==1.0
python-bitcoinlib==0.7.0
pytz==2017.2
requests==2.18.1
six==1.10.0
uritemplate==3.0.0
urllib3==1.21.1
